+++ packages/auth/src/config.ts
@@ -1,4 +1,8 @@
-import type { DefaultSession, NextAuthConfig } from "next-auth";
+import type {
+  DefaultSession,
+  NextAuthConfig,
+  Session as NextAuthSession,
+} from "next-auth";
 import { DrizzleAdapter } from "@auth/drizzle-adapter";
 import Discord from "next-auth/providers/discord";
 
@@ -13,12 +17,14 @@ declare module "next-auth" {
   }
 }
 
+const adapter = DrizzleAdapter(db, {
+  usersTable: User,
+  accountsTable: Account,
+  sessionsTable: Session,
+});
+
 export const authConfig = {
-  adapter: DrizzleAdapter(db, {
-    usersTable: User,
-    accountsTable: Account,
-    sessionsTable: Session,
-  }),
+  adapter,
   providers: [Discord],
   callbacks: {
     session: (opts) => {
@@ -34,3 +40,22 @@ export const authConfig = {
     },
   },
 } satisfies NextAuthConfig;
+
+export const validateToken = async (
+  token: string,
+): Promise<NextAuthSession | null> => {
+  const sessionToken = token.slice("Bearer ".length);
+  const session = await adapter.getSessionAndUser?.(sessionToken);
+  return session
+    ? {
+        user: {
+          ...session.user,
+        },
+        expires: session.session.expires.toISOString(),
+      }
+    : null;
+};
+
+export const invalidateSessionToken = async (token: string) => {
+  await adapter.deleteSession?.(token);
+};
