+++ packages/auth/src/index.ts
@@ -1,4 +1,3 @@
-import type { Session } from "next-auth";
 import NextAuth from "next-auth";
 
 import { authConfig } from "./config";
@@ -14,19 +13,4 @@ const {
 
 export { GET, POST, auth, signIn, signOut };
 
-export const validateToken = async (token: string): Promise<Session | null> => {
-  const sessionToken = token.slice("Bearer ".length);
-  const session = await authConfig.adapter.getSessionAndUser?.(sessionToken);
-  return session
-    ? {
-        user: {
-          ...session.user,
-        },
-        expires: session.session.expires.toISOString(),
-      }
-    : null;
-};
-
-export const invalidateSessionToken = async (token: string) => {
-  await authConfig.adapter.deleteSession?.(token);
-};
+export { invalidateSessionToken, validateToken } from "./config";
